task-classifier/requirements.txt
torch>=1.9.0
transformers>=4.18.0
datasets==2.13.1
scikit-learn>=1.0.2
tensorboard>=2.8.0
matplotlib==3.7.2
pandas==2.0.3
numpy==1.25.1
tqdm>=4.64.0
pytest==7.4.0
pyyaml>=6.0
fastapi>=0.85.0
uvicorn>=0.18.0
pydantic>=1.9.0
